Made new files functions.c/.h that contain all the neccessary functions except RunSimulation ,updated makefile and added option make clean to get rid of the objective files,updated README.md of the new installation proccess.My aim by making the new modules is to make it easier to read
