+++ functions.h
@@ -1,0 +1,12 @@
+#ifndef __FUNCTIONS__
+#define __FUNCTIONS___
+
+#define OP_NUM 5 /* Number of operators */
+#define RANGE 100 /* Define range of b */
+
+	double fitness(int  ,int );
+	char operate(long *,long );
+	int deviation(int ,int );
+
+
+#endif
